requirements.txt
# 自动询价系统依赖

# 核心
pyyaml>=6.0
requests>=2.28.0
httpx>=0.24.0
pydantic>=2.0

# Web 框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# 数据库
sqlalchemy>=2.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# 缓存
redis>=4.6.0

# 认证
PyJWT>=2.8.0

# 调度
apscheduler>=3.10.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0
flake8>=6.1.0

# 可选: 可视化
matplotlib>=3.7.0

# 可选: Excel 导出
openpyxl>=3.1.0

# 可选: 向量搜索 (提升历史匹配准确度)
# scikit-learn>=1.3.0
# numpy>=1.24.0

# 可选: 语义搜索 (本地嵌入模型)
# sentence-transformers>=2.2.0
